@@ -5,7 +5,7 @@
 
     <groupId>de.mindconsulting</groupId>
     <artifactId>ladon-s3-server-boot</artifactId>
-    <version>2.1.2</version>
+    <version>2.4.0</version>
     <packaging>jar</packaging>
 
     <name>s3-server-boot</name>
@@ -14,15 +14,16 @@
     <parent>
         <groupId>org.springframework.boot</groupId>
         <artifactId>spring-boot-starter-parent</artifactId>
-        <version>2.5.2</version>
+        <version>3.0.5</version>
         <relativePath/> <!-- lookup parent from repository -->
     </parent>
 
     <properties>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-        <java.version>1.8</java.version>
+        <java.version>17</java.version>
         <s3-server.version>${project.version}</s3-server.version>
         <dockerfile-maven-version>1.4.12</dockerfile-maven-version>
+        <spring.boot.version>3.0.5</spring.boot.version>
     </properties>
 
     <dependencies>
@@ -49,7 +50,7 @@
         <dependency>
             <groupId>com.amazonaws</groupId>
             <artifactId>aws-java-sdk-s3</artifactId>
-            <version>1.10.54</version>
+            <version>1.12.40</version>
             <scope>test</scope>
         </dependency>
     </dependencies>
@@ -73,6 +74,7 @@
             <plugin>
                 <groupId>org.springframework.boot</groupId>
                 <artifactId>spring-boot-maven-plugin</artifactId>
+                <version>3.0.5</version>
             </plugin>
 <!--            <plugin>-->
 <!--                <groupId>com.spotify</groupId>-->
