@@ -5,7 +5,7 @@
 
     <groupId>de.mindconsulting</groupId>
     <artifactId>ladon-s3-server</artifactId>
-    <version>2.1.2</version>
+    <version>2.4.0</version>
     <build>
         <plugins>
             <plugin>
@@ -19,6 +19,7 @@
             <plugin>
                 <groupId>org.apache.felix</groupId>
                 <artifactId>maven-bundle-plugin</artifactId>
+                <version>5.1.2</version>
                 <extensions>true</extensions>
                 <configuration>
                     <instructions>
@@ -27,10 +28,11 @@
                             de.mc.ladon.s3server.servlet,
                             de.mc.ladon.s3server.repository.api,
                             de.mc.ladon.s3server.entities.*,
+                            de.mc.ladon.s3server.entities.api.*,
                             de.mc.ladon.s3server.logging,
                             de.mc.ladon.s3server.exceptions,
                             de.mc.ladon.s3server.common,
-
+                            de.mc.ladon.s3server.auth,
                         </Export-Package>
                         <Private-Package>de.mc.ladon.s3server.*</Private-Package>
                         <Bundle-Name>Mind Consulting :: Ladon S3 Server</Bundle-Name>
@@ -50,7 +52,7 @@
 
     <properties>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-        <java.version>1.8</java.version>
+        <java.version>17</java.version>
     </properties>
 
     <dependencies>
@@ -58,12 +60,12 @@
         <dependency>
             <groupId>com.google.guava</groupId>
             <artifactId>guava</artifactId>
-            <version>19.0</version>
+            <version>31.1-jre</version>
         </dependency>
         <dependency>
-            <groupId>javax.xml.bind</groupId>
-            <artifactId>jaxb-api</artifactId>
-            <version>2.3.0</version>
+            <groupId>jakarta.xml.bind</groupId>
+            <artifactId>jakarta.xml.bind-api</artifactId>
+            <version>4.0.0</version>
         </dependency>
         <dependency>
             <groupId>org.apache.servicemix.bundles</groupId>
@@ -72,14 +74,14 @@
         </dependency>
 
         <dependency>
-            <groupId>javax.servlet</groupId>
-            <artifactId>javax.servlet-api</artifactId>
-            <version>3.1.0</version>
+            <groupId>jakarta.servlet</groupId>
+            <artifactId>jakarta.servlet-api</artifactId>
+            <version>6.0.0</version>
         </dependency>
         <dependency>
             <groupId>org.slf4j</groupId>
             <artifactId>slf4j-api</artifactId>
-            <version>1.7.18</version>
+            <version>2.0.7</version>
         </dependency>
         <dependency>
             <groupId>junit</groupId>
