@@ -5,7 +5,7 @@
 
     <groupId>de.mindconsulting</groupId>
     <artifactId>s3-server</artifactId>
-    <version>2.1.2</version>
+    <version>2.4.0</version>
     <packaging>pom</packaging>
 
     <name>s3-server-boot</name>
@@ -17,11 +17,12 @@
         <module>ladon-s3-server-fs</module>
         <module>ladon-s3-server-boot</module>
         <module>ladon-s3-server-osgi-starter</module>
+        <module>ladon-s3-server-jetty</module>
     </modules>
 
     <properties>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-        <java.version>1.8</java.version>
+        <java.version>17</java.version>
     </properties>
     
     <developers>
@@ -38,14 +39,7 @@
         <developerConnection>scm:git:git@github.com:mindmill/ladon-s3-server.git</developerConnection>
         <url>git@github.com:mindmill/ladon-s3-server.git</url>
     </scm>
-    <build>
-        <plugins>
-            <plugin>
-                <groupId>org.springframework.boot</groupId>
-                <artifactId>spring-boot-maven-plugin</artifactId>
-            </plugin>
-        </plugins>
-    </build>
+
 
     <distributionManagement>
         <repository>
