@@ -7,8 +7,8 @@ package de.mc.ladon.s3server.jaxb.fsmeta;
 
 import de.mc.ladon.s3server.repository.impl.FSUser;
 
-import javax.xml.bind.annotation.XmlElement;
-import javax.xml.bind.annotation.XmlRootElement;
+import jakarta.xml.bind.annotation.XmlElement;
+import jakarta.xml.bind.annotation.XmlRootElement;
 import java.util.Map;
 
 /**
