@@ -4,8 +4,8 @@
 
 package de.mc.ladon.s3server.jaxb.entities;
 
-import javax.xml.bind.annotation.XmlElement;
-import javax.xml.bind.annotation.XmlRootElement;
+import jakarta.xml.bind.annotation.XmlElement;
+import jakarta.xml.bind.annotation.XmlRootElement;
 
 /**
  * @author Ralf Ulrich on 27.02.16.
