@@ -5,6 +5,6 @@
 /**
  * @author Ralf Ulrich on 17.02.16.
  */
-@javax.xml.bind.annotation.XmlSchema(namespace = "http://s3.amazonaws.com/doc/2006-03-01/", elementFormDefault = javax.xml.bind.annotation.XmlNsForm.QUALIFIED, xmlns = { @javax.xml.bind.annotation.XmlNs(namespaceURI = "http://s3.amazonaws.com/doc/2006-03-01/", prefix = "") })
+@jakarta.xml.bind.annotation.XmlSchema(namespace = "http://s3.amazonaws.com/doc/2006-03-01/", elementFormDefault = jakarta.xml.bind.annotation.XmlNsForm.QUALIFIED, xmlns = { @jakarta.xml.bind.annotation.XmlNs(namespaceURI = "http://s3.amazonaws.com/doc/2006-03-01/", prefix = "") })
 package de.mc.ladon.s3server.jaxb.entities;
 
